requirements.txt
numpy==1.18.1
torch>=1.2.0
torchvision>=0.4.0
Pillow>=7.1.1
torchlars==0.1.2
tqdm==4.45.0
